requirements.txt
pandas
requests
python-dotenv
faker
